Moves transaction manager interface
