+++ src/core/interfaces/managers/TransactionManagerInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace DrlArchive\core\interfaces\repositories;
+namespace DrlArchive\core\interfaces\managers;
 
 
 interface TransactionManagerInterface
