+++ tests/mocks/TransactionManagerSpy.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace DrlArchive\mocks;
 
 
-use DrlArchive\core\interfaces\repositories\TransactionManagerInterface;
+use DrlArchive\core\interfaces\managers\TransactionManagerInterface;
 
 class TransactionManagerSpy implements TransactionManagerInterface
 {
